fix: links to plaintext
current snaps-ui does not support
markdown linking. to keep links in scope
of the audit, they will be kept as
plaintext.
also "Contract Tag" when not found is
now correctly in bold.
also don't display "Is this a token?" as
it would be annoying to have it in all
non-token contracts