fix: contains -> starts_with + ends_with
"contains_nocase" is a problematic
match check. in the future, new
namespaces would potentially be of
the form "${thing}eip155:1:0xc0fec0fe..."
that would cause matches, despite not
belonging to the same namespace.

the reason "contains" was used was,
because there's no simple check with
"nocase".
if the subgraph tracked the length of
keyN strings, we could assert nocase
identity with a combined query of
"starts with" and length.
but, the way CAIP-10 strings work,
a double assertion of starts_with
and ends_with will always function
as intended, as long as the registry
rejects bad submissions (which is trusted to do)

fix: #35