chore: use URL searchParams api
even though there are reasons this might not be needed,
it's good practice. kept as comments for later integration.
the raw hex `contractAddress` argument for `getInsights`
turns out to not be needed.
it might have slipped through when testing with earlier registry
versions. it won't be needed when links are available.

fix #37